Desktop/kinro/pyproject.toml
[tool.poetry]
name = "kinro"
version = "0.1.0"
description = "Un sistema de modelos dinámicos con FastAPI y React"
authors = ["Tu Nombre <hugo@hfmworks.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.27"
asyncpg = "^0.29.0"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"
alembic = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
